+++ src/main/java/org/folio/tm/integration/keycloak/configuration/KeycloakSessionProperties.java
@@ -1,0 +1,12 @@
+package org.folio.tm.integration.keycloak.configuration;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Data
+@ConfigurationProperties
+public class KeycloakSessionProperties {
+
+  private Integer idleTimeout;
+  private Integer maxLifespan;
+}